Reposition AcceptOrDeny component
